Made changes that allow the code base to be successfully published and run at least on the home machine - needs to be tested on the college machines and perhaps a laptop to make sure it will work.
The font.xnb file had to be added to the content folder and have the build action type changed to content rather than none.

Changes were also made to the default graph size and the speed of the robot to set it to a more appropriate size for the new grid.

The camera was also changed to not allow zooming in further than the default zoom level of 1.0f this is because there is no need for a greater zoom.