Removed the unnecessary RobotScanner class - the design has changed since the file was created and was no longer needed.
Began implementing the breadth first search algorithm with minimal luck so far. Need to speak to one of the tutors about potential solutions to the problem.

A change was made to the Cell class that allows the parent position vector to be nullable - this was done by changing the type from Vector2 to Vector2? which makes it a nullable type.

Performed general code clean up and comment clean up to remove unnecessary extras.