Added a method to the Graph to return a path from any point back to the root node.
Removed the reliance on the bool moving value and now solely relies on the count property of the movement queue instead.

Created a new solution to the breadth first search problem. The robot is now correctly able to find its way around the grid using this algorithm. The solution was to find a common intersection between the current position and the root node and the next parent position and the root node.

Currently there is one issue that revolves around the fact that the forward and return moves are added at the same time. This means that the last node will always be the one before the final explored node. This might not be a problem as the robot moves back to the start position after this - however there is a momentary delay in doing so.