Created a drawing interface to abstract the drawing code away from the main Game1.cs file. This allowed the creation of a class for drawing the grid, the robot and the user interface.
Moved the find unvisited neighbours code out of the robot class and into the Graph class to be more consistent with design.

Created an abstract search class and created separate classes for each of the search types in effort to reduce clutter in the main robot class. This also meant removing the search mode enum and replacing it with a field for search type.

The active field in the robot class was also change to be an instance of an enumeration called RobotStatus - this allows the use of idle, active and complete statuses for use in the simulation.

Added a simulation class that will be used to monitor the performance of each of the algorithms, so far it controls the starting and stopping of the robot and a stopwatch to measure the time taken to complete the exploration.

Finally completed the code for breadth first searching - still one minor bug persisting but this should not be a huge problem, it might be ignorable.

Began working on the A* implementation to provide help to the BFS and DFS heuristic based searches. There are currently some bugs with this and needs to be looked at. Not entirely sure where the problem lies.