The code has been changed to allow the simulation class a lot more control over the running of the tests. It is now possible to repeat tests and also perform the same test for each different search type.
The grid generation has had a seed added so that the tests can be peformed on the same grids and will effectively show consistent results.

The simulation will now output the results of the test to a CSV file ready for analysis.

The ordering of directions has been changed to North, West, South and East. This will effect any further recordings of the simulations.

The search algorithms have been completed with the A* algorithm being separated into it's own class file. The depth first search algorithm has also been implemented and tested.

Initial results from the simulations show what was expected - that the depth first a* algorithm is the fastest and covers the least number of tiles in each of the performed tests.